chore: remove RFC JSON from Nuxt app
