+++ .github/workflows/extractor-tests.yml
@@ -5,7 +5,6 @@ on:
     paths:
       - 'rfc-html-extractor/**'
       - 'client/app/utilities/rfc-validators.ts'
-      - 'client/app/utilities/rfc-converter-parse.ts'
   workflow_dispatch:
 
 jobs:
