+++ client/app/utilities/rfc-validators.ts
@@ -300,49 +300,3 @@ export const isNodePojo = (maybeNode: unknown): maybeNode is NodePojo => {
     ['Element', 'Text'].includes(maybeNode.type)
   )
 }
-
-const RfcJsonStatusSchema = z.union([
-  z.literal('UNKNOWN'),
-  z.literal('INTERNET STANDARD'),
-  z.literal('EXPERIMENTAL'),
-  z.literal('HISTORIC'),
-  z.literal('INFORMATIONAL'),
-  z.literal('NOT ISSUED'),
-  z.literal('PROPOSED STANDARD'),
-  z.literal('BEST CURRENT PRACTICE'),
-  z.literal('FOR YOUR INFORMATION')
-])
-
-const RfcJsonFormatSchema = z.union([
-  z.literal('ASCII'),
-  z.literal('TEXT'),
-  z.literal('HTML'),
-  z.literal('PDF'),
-  z.literal('PS'),
-  z.literal('XML'),
-  z.literal('')
-])
-
-export const RfcJsonSchema = z.object({
-  draft: z.string().nullable(),
-  doc_id: z.string(),
-  title: z.string(),
-  authors: z.array(z.string()),
-  format: RfcJsonFormatSchema.array(),
-  page_count: z.string().nullable(),
-  pub_status: RfcJsonStatusSchema,
-  status: RfcJsonStatusSchema,
-  source: z.string(),
-  abstract: z.string().nullable(),
-  pub_date: z.string().nullable(),
-  keywords: z.array(z.string()),
-  obsoletes: z.array(z.string()),
-  obsoleted_by: z.array(z.string()),
-  updates: z.array(z.string()),
-  updated_by: z.array(z.string()),
-  see_also: z.array(z.string()),
-  doi: z.string().nullable(),
-  errata_url: z.string().nullable()
-})
-
-export type RFCJSON = z.infer<typeof RfcJsonSchema>
