+++ rfc-html-extractor/src/utilities/rfc-converters-utils.ts
@@ -1,8 +1,5 @@
 import { assertNever } from './typescript.ts'
-import type {
-  RfcCommon,
-  RFCJSON
-} from '../../../client/app/utilities/rfc-validators.ts'
+import type {  RfcCommon } from '../../../client/app/utilities/rfc-validators.ts'
 import { DateTime } from 'luxon'
 
 type RfcAuthor = RfcCommon['authors'][number]
@@ -104,52 +101,4 @@ export const formatDatePublished = (
     return dt.toFormat('d LLLL yyyy')
   }
   return dt.toFormat('LLLL yyyy')
-}
-
-export const formatRfcStatusAsRfcJsonStatus = (
-  status: RfcCommon['status']
-): RFCJSON['status'] => {
-  switch (status.name) {
-    case 'Best Current Practice':
-      return 'BEST CURRENT PRACTICE'
-    case 'Draft Standard':
-      return 'PROPOSED STANDARD'
-    case 'Experimental':
-      return 'EXPERIMENTAL'
-    case 'Historic':
-      return 'HISTORIC'
-    case 'Informational':
-      return 'INFORMATIONAL'
-    case 'Internet Standard':
-      return 'INTERNET STANDARD'
-    case 'Not Issued':
-      return 'NOT ISSUED'
-    case 'Proposed Standard':
-      return 'PROPOSED STANDARD'
-    case 'Unknown':
-      return 'UNKNOWN'
-    case 'FYI':
-      return 'FOR YOUR INFORMATION'
-  }
-  assertNever(status)
-}
-
-export const formatRfcFormatAsRfcJsonFormat = (
-  format: RfcCommon['formats'][number]
-): RFCJSON['format'][number] => {
-  switch (format) {
-    case 'txt':
-      return 'TEXT'
-    case 'xml':
-      return 'XML'
-    case 'html':
-      return 'HTML'
-    case 'htmlized':
-      return 'HTML'
-    case 'pdf':
-      return 'PDF'
-    case 'ps':
-      return 'PS'
-  }
-  assertNever(format)
 }
